requirements.txt
djangorestframework
Django==1.11.6
drfdocs